webapp/requirements.txt
elasticsearch==7.13.0
python-dotenv==0.21.0
streamlit==1.12.2
